moved highligtText function to helper file
